Move openPlayLists function to useSpotify hook
